Type Checking

+++ src/commands/join.ts
@@ -30,7 +30,9 @@ export async function run(interaction: CommandInteraction) {
       interaction
         .guild!.channels.fetch(plan.channelId)
         .then(async (channel) => {
-          channel!.messages
+          if (channel === null || !channel.isText()) return;
+
+          channel.messages
             .fetch(plan.messageId)
             .then(async (message) => {
               await message.delete();
@@ -51,7 +53,8 @@ export async function run(interaction: CommandInteraction) {
 
       // Save Last Message
       interaction.fetchReply().then(async (message) => {
-        console.log(message);
+        if (!("channelId" in message)) return;
+
         await data.lastMessage(message.channelId, message.id);
       });
     } else {
